+++ tests/common.py
@@ -12,7 +12,6 @@ import numpy as np
 import pandas as pd
 
 import dios
-from saqc.constants import BAD
 from saqc.core import Flags
 from saqc.core.history import History, createHistoryFromData